Rename/reorganize the NotificationModel
